+++ shared/models/interfaces/common.py
@@ -1,0 +1,9 @@
+"""Provide comon tools for Dragos interfaces."""
+
+from pydantic import BaseModel, ConfigDict
+
+
+class FrozenBaseModel(BaseModel):
+    """Base class for frozen models. I.e Not alter-able after model_post_init."""
+
+    model_config = ConfigDict(frozen=True)
