+++ shared/models/octi/enum.py
