+++ shared/models/octi/observables/__init__.py
